modification type prix

+++ booking-api/src/main/java/ca/ulaval/glo4002/booking/api/resources/profit/dto/ProfitResponse.java
@@ -3,12 +3,12 @@ package ca.ulaval.glo4002.booking.api.resources.profit.dto;
 import com.fasterxml.jackson.annotation.JsonCreator;
 
 public class ProfitResponse {
-    public final double in;
-    public final double out;
-    public final double profit;
+    public final float in;
+    public final float out;
+    public final float profit;
      
     @JsonCreator
-    public ProfitResponse(double income, double outcome, double profit) {
+    public ProfitResponse(float income, float outcome, float profit) {
         this.in = income;
         this.out = outcome;
         this.profit = profit;
